moved serialization into own project
