+++ src/app/src/main/kotlin/org/vaccineimpact/api/app/controllers/endpoints/StreamedEndpoint.kt
@@ -1,8 +1,8 @@
 package org.vaccineimpact.api.app.controllers.endpoints
 
 import org.vaccineimpact.api.app.ActionContext
-import org.vaccineimpact.api.app.serialization.Serializer
-import org.vaccineimpact.api.app.serialization.StreamSerializable
+import org.vaccineimpact.api.serialization.Serializer
+import org.vaccineimpact.api.serialization.StreamSerializable
 
 fun <TRepository> ((ActionContext, TRepository) -> StreamSerializable<*>).streamed()
         : (ActionContext, TRepository) -> Unit
