+++ src/app/src/main/kotlin/org/vaccineimpact/api/app/repositories/jooq/JooqRepository.kt
@@ -1,8 +1,8 @@
 package org.vaccineimpact.api.app.repositories.jooq
 
 import org.jooq.DSLContext
-import org.vaccineimpact.api.Deserializer
-import org.vaccineimpact.api.UnknownEnumValue
+import org.vaccineimpact.api.serialization.Deserializer
+import org.vaccineimpact.api.serialization.UnknownEnumValue
 import org.vaccineimpact.api.app.errors.BadDatabaseConstant
 import org.vaccineimpact.api.app.repositories.Repository
