+++ src/app/src/main/kotlin/org/vaccineimpact/api/app/security/TokenIssuingConfigFactory.kt
@@ -5,9 +5,8 @@ import org.pac4j.core.config.ConfigFactory
 import org.pac4j.core.context.HttpConstants
 import org.pac4j.http.client.direct.DirectBasicAuthClient
 import org.pac4j.sparkjava.SparkWebContext
-import org.vaccineimpact.api.app.repositories.AccessLogRepository
 import org.vaccineimpact.api.app.repositories.RepositoryFactory
-import org.vaccineimpact.api.app.serialization.Serializer
+import org.vaccineimpact.api.serialization.Serializer
 import org.vaccineimpact.api.models.FailedAuthentication
 import spark.Spark as spk
