--- src/databaseInterface/src/main/kotlin/org/vaccineimpact/api/ContentTypes.kt
@@ -1,7 +1,0 @@
-package org.vaccineimpact.api
-
-object ContentTypes
-{
-    val csv = "text/csv"
-    val json = "application/json"
-}
