--- src/databaseInterface/src/main/kotlin/org/vaccineimpact/api/OneTimeAction.kt
@@ -1,9 +1,0 @@
-package org.vaccineimpact.api
-
-enum class OneTimeAction
-{
-    COVERAGE,
-    DEMOGRAPHY,
-    SET_PASSWORD,
-    BURDENS
-}
