+++ src/serialization/src/main/kotlin/org/vaccineimpact/api/serialization/RuleBasedSerializer.kt
@@ -1,4 +1,4 @@
-package org.vaccineimpact.api.app.serialization
+package org.vaccineimpact.api.serialization
 
 import com.github.salomonbrys.kotson.SerializerArg
 import com.github.salomonbrys.kotson.jsonSerializer
