+++ src/serialization/src/main/kotlin/org/vaccineimpact/api/serialization/UnknownEnumValue.kt
@@ -1,4 +1,4 @@
-package org.vaccineimpact.api
+package org.vaccineimpact.api.serialization
 
 class UnknownEnumValue(val name: String, val type: String)
     : Exception("Unable to parse '$name' as enum of type $type")
