+++ src/serialization/src/main/kotlin/org/vaccineimpact/api/serialization/validation/ValidationException.kt
@@ -1,0 +1,6 @@
+package org.vaccineimpact.api.serialization.validation
+
+import org.vaccineimpact.api.models.ErrorInfo
+
+class ValidationException(val errors: List<ErrorInfo>)
+    : Exception()
