+++ src/serialization/src/test/FlexibleDataTableTests.kt
@@ -2,8 +2,8 @@ package org.vaccineimpact.api.tests.serialization
 
 import org.assertj.core.api.Assertions
 import org.junit.Test
-import org.vaccineimpact.api.app.serialization.FlexibleDataTable
-import org.vaccineimpact.api.app.serialization.Serializer
+import org.vaccineimpact.api.serialization.FlexibleDataTable
+import org.vaccineimpact.api.serialization.Serializer
 import org.vaccineimpact.api.models.helpers.FlexibleProperty
 import org.vaccineimpact.api.test_helpers.MontaguTests
 import org.vaccineimpact.api.test_helpers.serializeToStreamAndGetAsString
