+++ src/serialization/src/test/SplitDataTests.kt
@@ -1,14 +1,11 @@
 package org.vaccineimpact.api.tests.serialization
 
-import com.nhaarman.mockito_kotlin.any
-import com.nhaarman.mockito_kotlin.doAnswer
-import com.nhaarman.mockito_kotlin.doReturn
 import com.nhaarman.mockito_kotlin.mock
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.Test
-import org.vaccineimpact.api.app.serialization.Serializer
-import org.vaccineimpact.api.app.serialization.SplitData
-import org.vaccineimpact.api.app.serialization.StreamSerializable
+import org.vaccineimpact.api.serialization.Serializer
+import org.vaccineimpact.api.serialization.SplitData
+import org.vaccineimpact.api.serialization.StreamSerializable
 import org.vaccineimpact.api.test_helpers.MontaguTests
 import org.vaccineimpact.api.test_helpers.serializeToStreamAndGetAsString
 import java.io.OutputStream
