+++ src/settings.gradle
@@ -8,6 +8,7 @@ include 'app',
         'generateDatabaseInterface',
         'generateTestData',
         'userCLI',
+        'serialization',
 
         'validateSchema',
         'testHelpers',
